Add hook to read/write pagination in location
The hook takes care of parsing the pagination from the location search
and provides a function that lets a consumer update the location with
the given pagination.
